+++ src/main/kotlin/org/rncp/ad/infra/db/AdDao.kt
@@ -13,12 +13,12 @@ data class AdDao(
         var userId: String,
         var name: String,
         var description: String,
-        var hourPrice: Float,
-        var latitude: Float,
-        var longitude: Float,
+        var hourPrice: Double,
+        var latitude: Double,
+        var longitude: Double,
         var state: Boolean,
 ) : PanacheEntityBase() {
-    constructor() : this(0, "", "", "", 0.0f, 0.0f, 0.0f, false)
+    constructor() : this(0, "", "", "", 0.0, 0.0, 0.0, false)
 
     fun toAd(): Ad {
         return Ad(id, userId, name, description, hourPrice, latitude, longitude, state)
